refactor: restructure component for chat detail section
