+++ src/screens/chat/ChatScreen.tsx
@@ -1,5 +1,5 @@
 import MyChatSection from "./sections/my-chat";
-import ChatDetailSection from "./sections/ChatDetailSection";
+import ChatDetailSection from "./sections/chat-detail";
 import { Fragment } from "react";
 
 const ChatScreen = () => {
