+++ src/screens/chat/sections/chat-detail/components/ChatAiIcon.tsx
