+++ src/screens/chat/sections/chat-detail/components/ChatProfileIcon.tsx
