+++ src/screens/chat/sections/chat-detail/index.ts
@@ -1,0 +1,3 @@
+import ChatDetailSection from "./ChatDetailSection";
+
+export default ChatDetailSection;
